Rebase and make requested changes
